Shopkeeper bug fixes

+++ src/main/java/net/dirtlands/listeners/shopkeepers/NPCDeletion.java
@@ -1,6 +1,7 @@
 package net.dirtlands.listeners.shopkeepers;
 
 import dirtlands.db.Tables;
+import net.citizensnpcs.api.event.DespawnReason;
 import net.citizensnpcs.api.event.NPCDespawnEvent;
 import net.dirtlands.Main;
 import org.bukkit.event.EventHandler;
@@ -10,8 +11,10 @@ public class NPCDeletion implements Listener {
 
     @EventHandler
     public void onNPCDespawn(NPCDespawnEvent e) {
-        Main.getPlugin().getDslContext().deleteFrom(Tables.SHOPKEEPERS)
-                .where(Tables.SHOPKEEPERS.SHOPKEEPERID.eq(e.getNPC().getId())).execute();
+        if (e.getReason().equals(DespawnReason.REMOVAL)) {
+            Main.getPlugin().getDslContext().deleteFrom(Tables.SHOPKEEPERS)
+                    .where(Tables.SHOPKEEPERS.SHOPKEEPERID.eq(e.getNPC().getId())).execute();
+        }
 
     }
